[RFC] Implement c10::Synchronized<T>, a basic error-avoiding synchronization wrapper for data
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73407

After adding mutex/lock in a previous change, I realized that `folly::Synchronized<T>` exists for this reason. However, we can't use it, so I have implemented a basic version for use in `c10` in this change.

Reviewers, please check if this looks reasonable as a minimal first step to enforce that data that is supposed to be used in a multi-threaded context ends up being locked before use. i.e. don't rely on the developer to do the right thing, instead enforce it using this abstraction.
ghstack-source-id: 149938601

Differential Revision: [D34465420](https://our.internmc.facebook.com/intern/diff/D34465420/)