[PyTorch] [Model Tracer] Use c10::Synchronized<T> abstraction for mutex protected data
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73408

Use the newly introduced `c10::Synchronized<T>` abstraction to enforce the use use of a mutex/lock on variables that should be protected for concurrent reads/modifications.

If reviewers think that this change is okay, I'll go ahead and change the other users in model tracer. This change is meant to show reviewers what the new code will look like.
ghstack-source-id: 149938602

Differential Revision: [D34465419](https://our.internmc.facebook.com/intern/diff/D34465419/)